requirements.txt
lark-parser==0.6.6
mido==1.2.9
python-rtmidi==1.2.1
urwid==2.0.1
